Fixes panics found in Juju for nil errors.
When supplying a nil error to an error type wrapper in Juju you risk a
panic on the Locationer as it relies on the error being non nil for the
Error() func to work.

Fix here is just to make an empty error when the error provided is nil.

This fix will allow users to not shoot their foot off.