Adds a shush method to the library and IsType.
Hide is responsible for making an error silent in fmt output.
Specifically useful with fmt.Errorf when you want an error to be wrapped
but want it's error message shushed.

IsType is a convenience method for checking a more complex error is of
type x within it's change. This is for cases where errors.Is can't be
used and where errors.As is not needed for the actual error type.